==> notebooks/00-symbols.ipynb <==
# Symbols

Das Projekt umfasst eine umfangreiche Codebasis. Um die Lesbarkeit und eine klare Orientierung in den Notebooks sicherzustellen, werden im Folgenden Symbole eingeführt, die als Hilfsmittel dienen sollen.

#### Rechenzeit

Im Rahmen der Notebooks werden teilweise rechenintensive Aufgaben durchgeführt. Die Ausführungsdauer dieser Aufgaben hängt in der Regel von der Rechenleistung des verwendeten Computers ab. Dennoch kann eine grobe Schätzung der Dauer wie folgt vorgenommen werden:

<font size="4">☕</font>: 5min - 20min

<font size="4">🎬</font>: 30min - 3h

<font size="4">🌒</font>: >> 12h

#### Weitere

<font size="4">🧩</font>: Zusammenfassender Codeblock. Führt notwendige Schritte in einem durch. Alternativ ist eine manuelle Durchführung der einzelnen Schritte unter dem Codeblock angegeben.

<font size="4">⭕️</font>: Es wird empfohlen, die Ausführung dieser Codeblöcke aufgrund ihrer sehr langen Laufzeit zu vermeiden. Die Ergebnisse des Codes wurden zuvor bereits in externen Dateien gesichert. In nachfolgenden Codeblöcken werden die Ergebnisse aus den Dateien geladen.
