==> notebooks/02-Data-Collection.ipynb <==
# Data Collection

Um später aussagekräftige Modelle erstellen zu können, müssen im ersten Schritt die dafür notwendigen Rohdaten gesammelt und ein Datensatz erstellt werden. Dazu werden Tweets von der Plattform Twitter gesammelt und in einer Liste gespeichert.

#### 1. Daten sammeln ⭕️🌒

Um ausschließlich relevante und die Branche betreffende Tweets zu sammeln werden dem Web-Scraper folgende Suchparameter übergeben:
- hashtags: #technology, #tech, #innovation
- language: en
- period: since:2018-04-01 until:2023-04-01
- limit: 1_000_000

# Hinweis: Der Web-Scraper kann nicht in einem Jupyter Notebook ausgeführt werden.
# Um den Web-Scraper auszuführen muss das Skript nitter_scraper_standalone_v2.py über ein Command-line Interface aufgerufen werden.

#### 2. Datensatz erstellen

Nachdem die Twitter-Beiträge gesammelt wurden, werden diese in ein Dataframe transformiert.

from src.data.twitter_scraper_standalone import *
import pandas as pd

# transform
df = transform_to_dataframe('../data/raw/data_1687865107.pkl')

# save raw dataframe
df.to_feather('../data/raw/twitter_tweets_raw.feather')

#### 3. Datensatz anzeigen

# set the maximum width of the columns to unlimited
pd.set_option('display.max_colwidth', None)

df.head(6)

---
